# back/docs/swagger.yaml
definitions:
  models.Door:
    properties:
      id:
        type: integer
      pathNode:
        $ref: '#/definitions/navigation.PathNode'
        type: object
      section:
        $ref: '#/definitions/models.Section'
        type: object
    type: object
  models.Room:
    properties:
      alias:
        items:
          type: string
        type: array
      color:
        type: string
      description:
        type: string
      doors:
        items:
          $ref: '#/definitions/models.Door'
        type: array
      floor:
        type: integer
      id:
        type: integer
      name:
        type: string
      pathNode:
        $ref: '#/definitions/navigation.PathNode'
        type: object
      sections:
        items:
          $ref: '#/definitions/models.Section'
        type: array
    type: object
  models.Section:
    properties:
      end:
        $ref: '#/definitions/navigation.Coordinate'
        type: object
      id:
        type: integer
      start:
        $ref: '#/definitions/navigation.Coordinate'
        type: object
    type: object
  navigation.Coordinate:
    properties:
      x:
        type: integer
      "y":
        type: integer
      z:
        type: integer
    type: object
  navigation.PathNode:
    properties:
      connectedNodes:
        items:
          $ref: '#/definitions/navigation.PathNode'
        type: array
      coordinate:
        $ref: '#/definitions/navigation.Coordinate'
        type: object
      group:
        $ref: '#/definitions/navigation.PathNodeGroup'
        type: object
      id:
        type: integer
    type: object
  navigation.PathNodeGroup:
    properties:
      id:
        type: integer
      name:
        type: string
      nodes:
        items:
          $ref: '#/definitions/navigation.PathNode'
        type: array
    type: object
info:
  contact: {}
  license: {}
paths:
  /navigation/dir:
    get:
      consumes:
      - application/json
      description: Gets the navigation route for a start and end room
      parameters:
      - description: the start room name
        in: query
        name: startroom
        type: string
      - description: the end room name
        in: query
        name: endroom
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/navigation.Coordinate'
            type: array
      summary: Get Navigation Route
      tags:
      - NavigationController
  /roomlist/:
    get:
      consumes:
      - application/json
      description: Gets all available rooms
      operationId: get-room-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Room'
            type: array
      summary: Get Room List
      tags:
      - RoomController
swagger: "2.0"